MB-23150 use CB-User-Agent to send cbq specific agent info
Since we had to add a CB-User-Agent header for QWB, I'm hacking godbc
to use that as well.
This has the advantage that we get the go http user agent version as well.
Also, now using the N1QL version, so as to know which version of cbq is
connecting.
As soon as I learn from Ceej how to add build number to the N1QL version,
we will have that added too.

Change-Id: I8793b5788f2052fe29ce6f3e65b9ed2011dbdd3f
Reviewed-on: http://review.couchbase.org/75542
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>