MB-70633 Changes for USING AI chat requests...
... sent via cbq shell.

Similar to the handling of transaction requests
sent via cbq shell. The chatid as well as the
server's address is noted and in the driver's connection
when a user requests to execute BEGIN AI CHAT statement.

This way the user need not worry about setting
natural_chatid parameter to make chat requests.
And chat requests are always sent to the same
server where it was created.

Noted information is set back to default, once
the user ends the chat by executing END AI CHAT.

Change-Id: I1cb07bbdccbb249d259af841b3fa29e4b9bba6c8
Reviewed-on: https://review.couchbase.org/c/godbc/+/227224
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>